EOS-6023 net/lnet: remove LNID cache (#71)
Currently we have a list of LNIDs, which is cached during m0tr.ko module
insertion. During the failover, LNID is added in runtime to LNet
configuration, which currently lifts the requirement to re-insert
m0tr.ko kernel module (and hence, the restart of all the Motr-processes
running on the node as well). This slows down the failover/failback time
considerably.

Solution: get rid of the LNIDs caching in m0tr.ko kernel module.
Instead, LNIDs are queried from LNET every time they are needed.

Signed-off-by: Maksym Medvied <max.medved@seagate.com>