Fix strncpy() and strncat() related issues
It is pointless to pass source string length to strncpy() or strncat().
It should be size of the destination buffer (minus 1).

strncpy() may produce truncated string and it is NOT null-terminated in
this case. Therefore, strncpy() is not a safe function and the
destination string should be terminated explicitly. In places which
aren't performance critical, it is enough to use snprintf() which is
safe.

Another issue is when strncat() is supplied with the destination buffer
size instead of maximum allowed suffix size. As result, it may lead to
a buffer overflow.